requirements.txt
streamlit
tensorflow
pillow
numpy
opencv-python
python-dotenv